# api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo API",
    "version": "v1"
  },
  "paths": {
...
    "/api/course": {
      "get": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseListDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseListDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseListDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Course"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/course/{id}": {
      "get": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponseDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/majors-category/{id}": {
      "get": {
        "tags": [
          "MajorsCategory"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MajorsCategoryResDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MajorsCategoryResDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MajorsCategoryResDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/majors-category": {
      "get": {
        "tags": [
          "MajorsCategory"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MajorsCategoryResDtoListDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MajorsCategoryResDtoListDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MajorsCategoryResDtoListDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MajorsCategory"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MajorsCategory"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MajorsCategory"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MajorsCategoryReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseRequestDto": {
        "required": [
          "courseTimes",
          "location",
          "name",
          "teacherId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "courseTimes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseTimeReqDto"
            }
          },
          "location": {
            "minLength": 1,
            "type": "string"
          },
          "teacherId": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CourseResponseDto": {
        "required": [
          "location",
          "name",
          "studentIds",
          "teacherId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "courseTimes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseTimeResDto"
            },
            "nullable": true
          },
          "location": {
            "minLength": 1,
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "teacherId": {
            "minLength": 1,
            "type": "string"
          },
          "studentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "CourseResponseDtoApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/CourseResponseDto"
          }
        },
        "additionalProperties": false
      },
      "CourseResponseListDto": {
        "type": "object",
        "properties": {
          "dataList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseResponseDto"
            },
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CourseResponseListDtoApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/CourseResponseListDto"
          }
        },
        "additionalProperties": false
      },
...
      "CourseTimeReqDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "timeTableId": {
            "type": "integer",
            "format": "int32"
          },
          "dateDay": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CourseTimeResDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "timeTableId": {
            "type": "integer",
            "format": "int32"
          },
          "dateDay": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "MajorsCategoryReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MajorsCategoryResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MajorsCategoryResDtoApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/MajorsCategoryResDto"
          }
        },
        "additionalProperties": false
      },
      "MajorsCategoryResDtoListDto": {
        "type": "object",
        "properties": {
          "dataList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MajorsCategoryResDto"
            },
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MajorsCategoryResDtoListDtoApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/MajorsCategoryResDtoListDto"
          }
        },
        "additionalProperties": false
      },
...
      "ObjectApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}